+++ lib/octantis_web/storybook/story.ex
@@ -16,6 +16,29 @@ defmodule OctantisWeb.Storybook.Story do
       def template do
         """
         <.stylesheet psb-code-hidden/>
+
+        <.psb-variation-group/>
+        """
+      end
+
+      defoverridable imports: 0, template: 0
+    end
+  end
+
+  def web_component do
+    quote do
+      use PhoenixStorybook.Story, :component
+
+      def imports,
+        do: [
+          {OctantisWeb.Components.Head, stylesheet: 1},
+          {OctantisWeb.Components.Head, javascript: 1}
+        ]
+
+      def template do
+        """
+        <.stylesheet psb-code-hidden/>
+        <.javascript psb-code-hidden/>
         <.psb-variation-group/>
         """
       end